Favicon, medias estrellas + progreso en %, admin desactivar/eliminar, puntuaciones del club y edición de fondos
- Favicon nuevo (huella 🐾) + icono para marcadores de iPhone — no existía ninguno real.
- Mi estantería: rating con medias estrellas (backend ya lo soportaba, solo
  faltaba en la UI) y progreso editable directamente por porcentaje además
  de por páginas, sincronizados entre sí.
- Admin puede desactivar (pausa reversible, echa al instante si tenía sesión
  abierta, conserva su estantería) y eliminar (borra todo) cuentas de otros
  jugadores — nueva pestaña "Desactivados" en el panel Admin.
- Puntuaciones abiertas de libros leídos por el club (reutiliza el modelo
  Vote existente, sin la parte de "secreto hasta revelar" que nunca tuvo
  interfaz) — cada uno pone/edita la suya con medias estrellas y ve las del
  resto desde la ficha del libro, con eco en el feed de Amigos.
- Los fondos de pantalla de Pirestore ya se pueden editar (nombre e imagen)
  sin borrar y recrear, conservando el mismo item_id para quien los tenga
  equipados.

+++ backend/database.py
@@ -143,7 +143,7 @@ class Vote(Base):
     id            = Column(Integer, primary_key=True, index=True)
     player_id     = Column(Integer, ForeignKey("players.id"), nullable=False)
     club_shelf_id = Column(Integer, ForeignKey("club_shelf.id"), nullable=False)
-    rating        = Column(Integer, nullable=False)   # 1-5
+    rating        = Column(Float, nullable=False)   # 0.5-5.0 con medios puntos
     revealed      = Column(Boolean, default=False)
     voted_at      = Column(DateTime, default=lambda: datetime.now(timezone.utc))
 
@@ -158,7 +158,7 @@ class Activity(Base):
     id         = Column(Integer, primary_key=True, index=True)
     player_id  = Column(Integer, ForeignKey("players.id", ondelete="CASCADE"), nullable=False)
     book_id    = Column(Integer, ForeignKey("books.id",   ondelete="CASCADE"), nullable=False)
-    event_type = Column(String, nullable=False)  # added | started | finished
+    event_type = Column(String, nullable=False)  # added | started | finished | proposed | voted
     rating     = Column(Float, nullable=True)
     created_at = Column(DateTime, default=lambda: datetime.now(timezone.utc))
